backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-dotenv==1.0.0

# PostgreSQL support (uncomment when switching to PostgreSQL)
# psycopg2-binary==2.9.9

